feat(ui): remove Profile page, consolidate into Settings
- new Wallets + API Access settings tabs (migrated from Profile)
- floating-glass Settings shell + dynamic per-tab header + header action
- AdminPanel legacy ThemedCard rewrite; SettingsPanels dialogs -> toast/confirm
- /profile -> /settings redirect
